idleRPG project
In a future not to long from here, we might start rewriting the existing slack-for-idleRPG code. 
The concept is old and frankly a bit boring (at least when we are less than 10 players). 
We're thinking something along the lines of; 
 - You start of as a peasant, progressing along a line of triggered event. 
 - Something bad might have happened and you are forced to leave your quiet peaceful mountain town to... um... save the princess?
 - You'll stop in villages and dungeons on the way performing certain actions.
 - Perhaps building up a base camp of your own with followers, sending your followers on missions to gather undefined resources? 
 - Teaming up with fellow players for certain fights?
 - BOSS FIGHTS!! Before you can move past certain points, you have to slay a monster. Or practice until you succeed in doing so.

Not sure if we should focus around Slack as a platform, or just build a hook in for simplifying sending commands. 

All in all, a little more RPG and less idle.